Moved HashableTracker to standalone file
